@@ -12,26 +12,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "angular-in-memory-web-api": "^0.2.5",
+    "@angular/animations": "^5.1.2",
+    "@angular/common": "^5.1.2",
+    "@angular/compiler": "^5.1.2",
+    "@angular/core": "^5.1.2",
+    "@angular/forms": "^5.1.2",
+    "@angular/http": "^5.1.2",
+    "@angular/platform-browser": "^5.1.2",
+    "@angular/platform-browser-dynamic": "^5.1.2",
+    "@angular/platform-server": "^5.1.2",
+    "@angular/router": "^5.1.2",
+    "angular-in-memory-web-api": "^0.5.2",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
     "include-media": "^1.4.9",
-    "rxjs": "^5.1.0",
+    "rxjs": "^5.5.6",
     "web-animations-js": "^2.2.5",
+    "webpack": "^3.10.0",
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
     "@angular/cli": "^1.4.2",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/compiler-cli": "^5.1.2",
     "@angular/language-service": "^4.0.0",
     "@types/greensock": "^1.15.32",
     "@types/gsap": "^1.19.1",
@@ -52,6 +54,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "2.4.2"
   }
 }
