@@ -1,7 +1,7 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HammerGestureConfig, HAMMER_GESTURE_CONFIG } from '@angular/platform-browser';
 
@@ -39,7 +39,7 @@ export class MyHammerConfig extends HammerGestureConfig {
   imports: [ 
   	BrowserModule, 
   	FormsModule, 
-  	HttpModule,
+  	HttpClientModule,
     InMemoryWebApiModule.forRoot(ProjectData, {delay: 0}),
     UserModule,
     MessageModule,
